FreezyLexer

Description

FreezyLexer - A Lexer for FreezyLang

Documentation

type Lexer a = ExceptT LexerError (StateT LexerState Identity) a

The Lexer monad stack

ExceptT for error handling

StateT for the lexerstate

Identity as the base monad

runLexer :: LexerState -> Lexer a -> (Either LexerError a, LexerState)

Unwrapping of the monad stack

data LexerState

The State of the Lexer

Constructors

LexerState

Fields

line :: Int

initState :: LexerState

initial state constructor

incLine :: Lexer ()

increments the line in the lexer state

queryLine :: Lexer Int

query the lexerstate for the current line

data LexerError

Type for lexer errors I'm not sure if this is really required. we can get the line from the state but i will stick with it for now

Constructors

LexerError

Fields

reason :: String

e_line :: Int

lexIt :: String -> Lexer [Token]

Tokenize the complete input

lexNext :: String -> Lexer (Token, String)

Lex the next token

lexComment :: String -> String -> Lexer (Token, String)

Lex a single line comment

lexNum :: String -> String -> Lexer (Token, String)

Lex an Integer number

lexIdent :: String -> String -> Lexer (Token, String)

Lex an identifier or reserved keyword

lexStr :: String -> String -> Lexer (Token, String)

Lex a string

Helper Functions

mkLitToken :: TokenType -> String -> String -> Lexer (Token, String)

convenience helper for literal tokens

mkToken :: TokenType -> String -> Lexer (Token, String)

convenience helper for tokens without an interesting lexeme

mkReservedOrIdentToken :: String -> String -> Lexer (Token, String)

convenience helper that maps the given lexeme to either a reserved keyword or an identifier

mkError :: String -> Lexer a

convenience helper to create an error on the current line

Char Helpers

isDigit :: Char -> Bool

simple wrapper around Data.Char.isDigit

isAlpha :: Char -> Bool

Data.Char.isAlpa or an underscore

isAlphaNum :: Char -> Bool

isDigit or isAlpha